paylist 부분 완전 수정

+++ src/main/java/com/example/cmtProject/dto/erp/salaries/PaymentDTO.java
@@ -1,5 +1,6 @@
 package com.example.cmtProject.dto.erp.salaries;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import lombok.AllArgsConstructor;
@@ -24,17 +25,17 @@ public class PaymentDTO {
     private LocalDate payDate;       // 지급일
 
     private Long payBasic;           // 기본급
-    private Double payBonusOvertime;   // 야근수당
-    private Double payBonusHoliday;    // 명절수당
+    private BigDecimal payBonusOvertime;   // 야근수당
+    private BigDecimal payBonusHoliday;    // 명절수당
 
     private Long payBonusTotal;      // 총수당금액
 
-    private Double payTaxPension;      // 국민연금
-    private Double payTaxCare;         // 장기요양보험
-    private Double payTaxHealth;       // 건강보험
-    private Double payTaxEmployment;   // 고용보험
-    private Double payTaxIncome;       // 소득세
-    private Double payTaxResidence;    // 주민세
+    private BigDecimal payTaxPension;      // 국민연금
+    private BigDecimal payTaxCare;         // 장기요양보험
+    private BigDecimal payTaxHealth;       // 건강보험
+    private BigDecimal payTaxEmployment;   // 고용보험
+    private BigDecimal payTaxIncome;       // 소득세
+    private BigDecimal payTaxResidence;    // 주민세
 
     private Long payTaxTotal;        // 총공제금액
 
@@ -44,5 +45,4 @@ public class PaymentDTO {
     
     private String salBankName;	     // 은행명
     private String salBankAccount;	 // 계좌번호
-    
 }
